backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic_settings
beautifulsoup4
requests
python-dotenv
aiohttp
httpx
